requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.44
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
pillow==10.1.0
python-multipart==0.0.6
cryptography==41.0.7